+++ src/lib.rs
@@ -45,8 +45,8 @@ pub use header::{IntoHeaderValue, XSpanIdString, X_SPAN_ID};
 #[cfg(feature = "multipart")]
 pub mod multipart;
 
-pub mod one_of;
-pub use one_of::*;
+pub mod one_any_of;
+pub use one_any_of::*;
 
 /// Helper Bound for Errors for MakeService/Service wrappers
 pub trait ErrorBound: Into<Box<dyn std::error::Error + Send + Sync>> {}
